refactor(#2412): split IndexReader's lazy-open primitives into index_reader/lazy.rs
Campsite #1116: index_reader.rs/mod.rs was already over the 800-line source
threshold before this issue; Stage 2's lazy-open additions (MaterializedIndex,
open_lazy, ensure_materialized + their unit tests) now live in a NEW submodule
(index_reader/lazy.rs, 219 lines) instead of growing mod.rs further. Residual
mod.rs growth (+7 lines net vs origin/main, from the doc-comment updates on the
now-OnceCell-backed get_partition_entries/is_fully_parsed/lookup_partition
accessors) and the pre-existing reader/mod.rs orchestration reorder (+9 lines,
summary-before-index load sequencing for open()) are genuinely inline with
those files' existing responsibilities -- not splittable without a much larger
out-of-scope refactor of those already-oversized epic #1116 backlog files.
Re-run with CQLITE_ALLOW_FILE_GROWTH=1 to acknowledge (note linked here).

Refs #2412 Stage 2 (design §A), epic #1116.
