+++ cqlite-core/src/storage/sstable/index_reader/mod.rs
@@ -134,41 +134,11 @@ pub struct IndexData {
     pub key_lookup: HashMap<Arc<[u8]>, usize>,
 }
 
-/// The full parse result, materialized either eagerly (at construction, the
-/// unchanged legacy behavior every direct [`IndexReader::open`] caller relies on)
-/// or lazily (issue #2412, on the first [`IndexReader::ensure_materialized`] call).
-struct MaterializedIndex {
-    index_data: IndexData,
-    /// Whether the entry parse consumed the ENTIRE Index.db file (issue #2302).
-    /// The parser `break`s on the first unparseable entry and returns the parsed
-    /// PREFIX, so a mid-entry-truncated file opens with leftover bytes. `true` ⟺
-    /// no bytes remained — the authoritative signal the stream was not cut
-    /// mid-entry (WHOLE trailing entries dropped at an exact boundary are caught
-    /// separately at the enumeration site). Only the completeness-sensitive full
-    /// enumeration consults it; point-lookup callers tolerate a partial prefix.
-    fully_parsed: bool,
-}
+// Lazy Summary-guided open (`open_lazy`/`ensure_materialized`, #2412 §A): see `lazy.rs`.
+mod lazy;
+use lazy::MaterializedIndex;
 
-/// High-level Index.db file reader
-///
-/// ## Lazy Summary-guided open (issue #2412)
-///
-/// [`Self::open`]/[`Self::open_with_summary`]/[`Self::open_with_summary_cancellable`]
-/// (the constructors every direct caller — including all pre-#2412 unit/integration
-/// tests — uses) are UNCHANGED: they parse the whole `Index.db` immediately and the
-/// [`MaterializedIndex`] cell is `set()` before the reader is returned, so
-/// [`Self::get_partition_entries`]/[`Self::lookup_partition`]/[`Self::is_fully_parsed`]
-/// behave EXACTLY as before for those callers.
-///
-/// [`Self::open_lazy`] is a NEW, additive constructor used only by
-/// `SSTableReader`'s BIG-open composition (`load_index_reader`) when a usable
-/// `Summary.db` is present: it performs zero `Index.db` I/O and leaves the cell
-/// unset. The first internal consumer that needs the full resident map calls
-/// [`Self::ensure_materialized`], which performs the identical full parse ONE time
-/// (memoized) — deferring the cost from open to first full-enumeration use, never
-/// eliminating it (Stage 3/4 of #2412 replace the common point-lookup/scan
-/// consumers with a bounded Summary-guided interval read that never touches this
-/// cell at all).
+/// High-level Index.db file reader. See `lazy.rs` for the #2412 lazy-open contract.
 #[allow(dead_code)]
 pub struct IndexReader {
     /// Path to the Index.db file
@@ -264,10 +234,8 @@ impl IndexReader {
         // complete. Structural, no heuristics (issue #28).
         let fully_parsed = remaining.is_empty();
 
+        // Eager: populate the cell immediately (fresh cell, `set` always succeeds).
         let materialized = tokio::sync::OnceCell::new();
-        // Fresh cell: `set` always succeeds. Eager construction populates it
-        // immediately so every pre-#2412 caller's sync accessors behave exactly as
-        // before this change.
         let _ = materialized.set(MaterializedIndex {
             index_data,
             fully_parsed,
@@ -280,78 +248,7 @@ impl IndexReader {
         })
     }
 
-    /// Lazy Summary-guided open (issue #2412, design §A): checks only that the file
-    /// EXISTS and performs ZERO `Index.db` parse work. Used by `SSTableReader`'s BIG
-    /// open composition when a usable `Summary.db` is present, so open cost is
-    /// bounded by `Summary.db` size, not by `Index.db` partition count. The full
-    /// parse is deferred to the first [`Self::ensure_materialized`] call from an
-    /// internal consumer that genuinely needs the resident map (a full-enumeration
-    /// scan, or the point-lookup fallback before Stage 3's interval-bounded lookup
-    /// lands); a point-lookup-dominated workload may never pay it at all.
-    pub(crate) async fn open_lazy(path: &Path, platform: Arc<Platform>) -> Result<Self> {
-        if !platform.fs().exists(path).await? {
-            return Err(Error::not_found(format!(
-                "Index.db file not found: {}",
-                path.display()
-            )));
-        }
-        Ok(Self {
-            file_path: path.to_path_buf(),
-            materialized: tokio::sync::OnceCell::new(),
-            platform,
-        })
-    }
-
-    /// Ensure the full `Index.db` parse has run, performing it exactly once
-    /// (memoized) on first call — the deferred cost [`Self::open_lazy`] skips at
-    /// open time (issue #2412). A no-op (immediate return) for a reader constructed
-    /// via the eager constructors, since their cell is already populated. Cancel-
-    /// aware: polls `cancel` the same way the eager parse does, so a cooperative
-    /// cancellation still aborts a large deferred parse promptly.
-    pub(crate) async fn ensure_materialized(
-        &self,
-        cancel: &crate::storage::scan_cancel::ScanCancel,
-    ) -> Result<()> {
-        self.materialized
-            .get_or_try_init(|| async {
-                cancel.check()?;
-                let mut file = File::open(&self.file_path).await?;
-                let mut buffer = Vec::new();
-                file.read_to_end(&mut buffer).await?;
-                if buffer.is_empty() {
-                    return Err(Error::corruption(format!(
-                        "Index.db file is empty: {}",
-                        self.file_path.display()
-                    )));
-                }
-                let (remaining, index_data) =
-                    match parse_index_data_cancellable(&buffer, None, cancel) {
-                        Ok(pair) => pair,
-                        Err(e @ Error::Cancelled) => return Err(e),
-                        Err(e) => {
-                            return Err(Error::corruption(format!(
-                                "Failed to parse Index.db: {:?}",
-                                e
-                            )));
-                        }
-                    };
-                let fully_parsed = remaining.is_empty();
-                Ok(MaterializedIndex {
-                    index_data,
-                    fully_parsed,
-                })
-            })
-            .await?;
-        Ok(())
-    }
-
-    /// Get all partition entries.
-    ///
-    /// Returns an empty slice if this reader was constructed via [`Self::open_lazy`]
-    /// and [`Self::ensure_materialized`] has not yet been called — every internal
-    /// consumer that needs the full map calls `ensure_materialized` first (issue
-    /// #2412). Constructors other than `open_lazy` always populate the cell before
-    /// returning, so this never returns a stale-empty slice for their callers.
+    /// Get all partition entries. Empty until materialized (see `lazy.rs`, #2412).
     pub fn get_partition_entries(&self) -> &[PartitionIndexEntry] {
         self.materialized
             .get()
@@ -360,10 +257,8 @@ impl IndexReader {
     }
 
     /// Whether the entry parse consumed the ENTIRE Index.db file (issue #2302).
-    /// `false` ⟺ the file was truncated mid-entry, so the parsed entry set is a
-    /// partial prefix and MUST NOT be treated as a complete partition enumeration.
-    /// Also `false` (conservatively) if not yet materialized (issue #2412) — every
-    /// full-enumeration consumer calls `ensure_materialized` first.
+    /// `false` ⟺ the file was truncated mid-entry — a partial prefix, not a
+    /// complete enumeration. Also `false` until materialized (`lazy.rs`, #2412).
     pub fn is_fully_parsed(&self) -> bool {
         self.materialized
             .get()
@@ -371,19 +266,8 @@ impl IndexReader {
             .unwrap_or(false)
     }
 
-    /// Look up a partition by key digest
-    ///
-    /// ## Zero-Allocation Optimization (Issue #107)
-    ///
-    /// This method performs HashMap lookup without heap allocation by leveraging
-    /// the `Borrow` trait. Since `Arc<[u8]>` implements `Borrow<[u8]>`, we can
-    /// lookup using `&[u8]` directly without creating a temporary Arc.
-    ///
-    /// **Before:** `let key_arc: Arc<[u8]> = key_digest.into();` (heap allocation per query)
-    /// **After:** Direct `get(key_digest)` using Borrow trait (zero allocations)
-    ///
-    /// Returns `None` if not yet materialized (issue #2412's lazy constructor) —
-    /// every internal caller ensures materialization first.
+    /// Look up a partition by key digest (zero-allocation `Borrow` lookup, Issue
+    /// #107). `None` until materialized (`lazy.rs`, #2412).
     pub fn lookup_partition(&self, key_digest: &[u8]) -> Option<&PartitionIndexEntry> {
         let m = self.materialized.get()?;
         m.index_data
@@ -459,130 +343,7 @@ mod tests {
     use super::*;
     use std::env;
 
-    /// Issue #2412 (Stage 2, §A): `open_lazy` performs ZERO `Index.db` parse work —
-    /// no entries touched, `is_fully_parsed()` conservatively `false` — until
-    /// `ensure_materialized` is called. Scale-free work probe: constructed from a
-    /// synthetic single-entry `Index.db` so the assertion is about ORDER of
-    /// operations, not entry count.
-    #[tokio::test]
-    async fn open_lazy_touches_zero_entries_until_materialized() {
-        let dir = std::env::temp_dir().join(format!(
-            "cqlite-2412-open-lazy-{}-{}",
-            std::process::id(),
-            std::time::SystemTime::now()
-                .duration_since(std::time::UNIX_EPOCH)
-                .map(|d| d.as_nanos())
-                .unwrap_or(0)
-        ));
-        std::fs::create_dir_all(&dir).unwrap();
-        let path = dir.join("nb-1-big-Index.db");
-        // One well-formed BIG entry: key_len(2) + key(4) + vint offset + vint promoted_len=0.
-        let entry: Vec<u8> = vec![0x00, 0x04, 0xAA, 0xBB, 0xCC, 0xDD, 0x00, 0x00];
-        std::fs::write(&path, &entry).unwrap();
-
-        let config = crate::Config::default();
-        let platform = Arc::new(
-            crate::Platform::new(&config)
-                .await
-                .expect("platform must initialize"),
-        );
-
-        let reader = IndexReader::open_lazy(&path, platform)
-            .await
-            .expect("open_lazy must succeed for an existing file");
-
-        // Zero work at open: no entries visible, not reported fully-parsed.
-        assert!(
-            reader.get_partition_entries().is_empty(),
-            "open_lazy must not materialize any entries before ensure_materialized"
-        );
-        assert!(
-            !reader.is_fully_parsed(),
-            "open_lazy must not report fully_parsed before materialization"
-        );
-        assert!(reader.lookup_partition(&[0xAA, 0xBB, 0xCC, 0xDD]).is_none());
-
-        // First real use triggers exactly the deferred full parse.
-        reader
-            .ensure_materialized(&crate::storage::scan_cancel::ScanCancel::default())
-            .await
-            .expect("deferred materialize must succeed for a well-formed file");
-        assert_eq!(reader.get_partition_entries().len(), 1);
-        assert!(reader.is_fully_parsed());
-        assert!(reader.lookup_partition(&[0xAA, 0xBB, 0xCC, 0xDD]).is_some());
-
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    /// `ensure_materialized` is idempotent/memoized: calling it twice does not
-    /// re-parse (the second call is a cheap `OnceCell::get` under the hood) and
-    /// yields the identical result both times.
-    #[tokio::test]
-    async fn ensure_materialized_is_memoized() {
-        let dir = std::env::temp_dir().join(format!(
-            "cqlite-2412-memoized-{}-{}",
-            std::process::id(),
-            std::time::SystemTime::now()
-                .duration_since(std::time::UNIX_EPOCH)
-                .map(|d| d.as_nanos())
-                .unwrap_or(0)
-        ));
-        std::fs::create_dir_all(&dir).unwrap();
-        let path = dir.join("nb-1-big-Index.db");
-        let entry: Vec<u8> = vec![0x00, 0x02, 0x11, 0x22, 0x00, 0x00];
-        std::fs::write(&path, &entry).unwrap();
-
-        let config = crate::Config::default();
-        let platform = Arc::new(
-            crate::Platform::new(&config)
-                .await
-                .expect("platform must initialize"),
-        );
-        let reader = IndexReader::open_lazy(&path, platform).await.unwrap();
-        let cancel = crate::storage::scan_cancel::ScanCancel::default();
-
-        reader.ensure_materialized(&cancel).await.unwrap();
-        let first_len = reader.get_partition_entries().len();
-        reader.ensure_materialized(&cancel).await.unwrap();
-        let second_len = reader.get_partition_entries().len();
-        assert_eq!(first_len, second_len);
-        assert_eq!(first_len, 1);
-
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    /// Eager constructors ([`IndexReader::open`]) are UNCHANGED: entries are
-    /// visible immediately, no `ensure_materialized` call needed — the pre-#2412
-    /// behavior every direct caller (including the many integration tests that
-    /// construct `IndexReader` directly) relies on.
-    #[tokio::test]
-    async fn eager_open_is_immediately_materialized() {
-        let dir = std::env::temp_dir().join(format!(
-            "cqlite-2412-eager-{}-{}",
-            std::process::id(),
-            std::time::SystemTime::now()
-                .duration_since(std::time::UNIX_EPOCH)
-                .map(|d| d.as_nanos())
-                .unwrap_or(0)
-        ));
-        std::fs::create_dir_all(&dir).unwrap();
-        let path = dir.join("nb-1-big-Index.db");
-        let entry: Vec<u8> = vec![0x00, 0x02, 0x33, 0x44, 0x00, 0x00];
-        std::fs::write(&path, &entry).unwrap();
-
-        let config = crate::Config::default();
-        let platform = Arc::new(
-            crate::Platform::new(&config)
-                .await
-                .expect("platform must initialize"),
-        );
-        let reader = IndexReader::open(&path, platform).await.unwrap();
-        assert_eq!(reader.get_partition_entries().len(), 1);
-        assert!(reader.is_fully_parsed());
-        assert!(reader.lookup_partition(&[0x33, 0x44]).is_some());
-
-        let _ = std::fs::remove_dir_all(&dir);
-    }
+    // #2412 lazy-open pins live in `lazy.rs`'s own test module (campsite #1116).
 
     /// Test stock_prices Index.db parsing (Issue #208)
     ///
@@ -762,10 +523,7 @@ mod tests {
 
                 // Check if index_reader was loaded (it's a public field)
                 if let Some(ref index_reader) = reader.index_reader {
-                    // Issue #2412: BIG open is now lazy (Summary-guided) when a
-                    // usable Summary.db is present, so the resident map is empty
-                    // until the first real full-enumeration/point-lookup demand.
-                    // Force materialization here to assert the on-disk entry count.
+                    // #2412: BIG open is lazy now; force materialization to assert count.
                     index_reader
                         .ensure_materialized(&crate::storage::scan_cancel::ScanCancel::default())
                         .await
